Changes to allow token-macro work with Run Vs AbstractBuild
