+++ core/src/main/java/hudson/model/BooleanParameterValue.java
@@ -63,7 +63,7 @@ public class BooleanParameterValue extends ParameterValue {
     }
 
     @Override
-    public VariableResolver<String> createVariableResolver(AbstractBuild<?, ?> build) {
+    public VariableResolver<String> createVariableResolver(Run<?, ?> build) {
         return new VariableResolver<String>() {
             public String resolve(String name) {
                 return BooleanParameterValue.this.name.equals(name) ? Boolean.toString(value) : null;
