+++ core/src/main/java/hudson/model/Run.java
@@ -58,6 +58,7 @@ import hudson.util.FlushProofOutputStream;
 import hudson.util.FormApply;
 import hudson.util.LogTaskListener;
 import hudson.util.ProcessTree;
+import hudson.util.VariableResolver;
 import hudson.util.XStream2;
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
@@ -348,7 +349,53 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
             artifactManager.onLoad(this);
         }
     }
-    
+
+    /**
+     * Creates {@link hudson.util.VariableResolver} backed by {@link #getBuildVariables()}.
+     */
+    public final VariableResolver<String> getBuildVariableResolver() {
+        return new VariableResolver.ByMap<String>(getBuildVariables());
+    }
+
+    /**
+     * Provides additional variables and their values to {@link hudson.tasks.Builder}s.
+     *
+     * <p>
+     * This mechanism is used by {@code MatrixConfiguration} to pass
+     * the configuration values to the current build. It is up to
+     * {@link hudson.tasks.Builder}s to decide whether they want to recognize the values
+     * or how to use them.
+     *
+     * <p>
+     * This also includes build parameters if a build is parameterized.
+     *
+     * @return
+     *      The returned map is mutable so that subtypes can put more values.
+     */
+    public Map<String,String> getBuildVariables() {
+        Map<String,String> r = new HashMap<String, String>();
+
+        ParametersAction parameters = getAction(ParametersAction.class);
+        if (parameters!=null) {
+            // this is a rather round about way of doing this...
+            for (ParameterValue p : parameters) {
+                String v = p.createVariableResolver(this).resolve(p.getName());
+                if (v!=null) r.put(p.getName(),v);
+            }
+        }
+
+        // allow the BuildWrappers to contribute additional build variables
+        if (project instanceof BuildableItemWithBuildWrappers) {
+            for (BuildWrapper bw : ((BuildableItemWithBuildWrappers) project).getBuildWrappersList())
+                bw.makeBuildVariables(this,r);
+        }
+
+        for (BuildVariableContributor bvc : BuildVariableContributor.all())
+            bvc.buildVariablesFor(this,r);
+
+        return r;
+    }
+
     /**
      * Return all transient actions associated with this build.
      *
