+++ core/src/main/java/hudson/model/StringParameterValue.java
@@ -58,7 +58,7 @@ public class StringParameterValue extends ParameterValue {
     }
 
     @Override
-    public VariableResolver<String> createVariableResolver(AbstractBuild<?, ?> build) {
+    public VariableResolver<String> createVariableResolver(Run<?, ?> build) {
         return new VariableResolver<String>() {
             public String resolve(String name) {
                 return StringParameterValue.this.name.equals(name) ? value : null;
